refactor : 대시보드 로컬 컴포넌트 파일 분리
* 추천 사정사 행·미니 카드·리포트 요약 카드를 형제 파일로 분리
* 대시보드 본문 분기를 DashboardHome 파일로 분리
* 이름만 같고 모양이 다른 두 단계 연결선을 TimelineConnector·StepArrow로 구분
* 진행 중 사건 행과 검수 대기 행을 형제 파일로 분리
