+++ apps/web/src/app/customer/dashboard/_components/DashboardView.tsx
@@ -1,11 +1,8 @@
 "use client";
 
-import { useDashboard } from "../_api/use-dashboard";
-import { DashboardDesktopView } from "./DashboardDesktopView";
+import { DashboardHome } from "./DashboardHome";
 import { SectionBoundary } from "./SectionBoundary";
 import { SectionSkeleton } from "./SectionSkeleton";
-import { MobileHomeView } from "./mobile/MobileHomeView";
-import { OnboardingHomeView } from "./onboarding/OnboardingHomeView";
 
 export function DashboardView() {
   return (
@@ -15,22 +12,5 @@ export function DashboardView() {
     >
       <DashboardHome />
     </SectionBoundary>
-  );
-}
-
-function DashboardHome() {
-  const { data } = useDashboard();
-
-  if (data.reportCount === 0) {
-    return <OnboardingHomeView />;
-  }
-
-  return (
-    <>
-      <div className="hidden md:block">
-        <DashboardDesktopView />
-      </div>
-      <MobileHomeView />
-    </>
   );
 }
