+++ apps/web/src/app/customer/dashboard/_components/onboarding/OnboardingSteps.tsx
@@ -1,6 +1,5 @@
-import type { ReactNode } from "react";
-import { ChevronRight } from "@/shared/ui/icons/ChevronRight";
 import { ONBOARDING_STEPS_ANCHOR } from "./onboarding-anchors";
+import { StepArrow } from "./StepArrow";
 
 interface OnboardingStep {
   order: string;
@@ -57,21 +56,10 @@ export function OnboardingSteps() {
               {step.description}
             </p>
 
-            {index < STEPS.length - 1 && <StepConnector />}
+            {index < STEPS.length - 1 && <StepArrow />}
           </li>
         ))}
       </ol>
     </section>
-  );
-}
-
-function StepConnector(): ReactNode {
-  return (
-    <span
-      aria-hidden
-      className="absolute -right-2 top-1/2 z-10 hidden size-[1.625rem] -translate-y-1/2 items-center justify-center rounded-input border border-line bg-paper-2 text-ink-3 md:flex"
-    >
-      <ChevronRight className="text-[0.8125rem]" />
-    </span>
   );
 }
